extract TeamSubscriptionOptions
